+++ cpp/api/src/base/Application.cpp
@@ -416,9 +416,9 @@ std::unique_ptr<ServerAndApp> This::connectToStarter(int options, int timeout) {
 	std::unique_ptr<App> app;
 
 	// Create the server with proxy or not.
-	bool useProxy = ((options & USE_PROXY) != 0);
+	bool useProxy = ((options & option::USE_PROXY) != 0);
 	if (useProxy) {
-		server = Server::create(m_instance.m_starterEndpoint.withPort(m_instance.m_starterProxyPort), USE_PROXY);
+		server = Server::create(m_instance.m_starterEndpoint.withPort(m_instance.m_starterProxyPort), option::USE_PROXY);
 	}
 	else {
 		server = Server::create(m_instance.m_starterEndpoint);
@@ -480,7 +480,7 @@ void This::initStarterCheck() {
 	// Create the starter server.
 	// If the starter has a running proxy, then use the proxy is reasonable.
 	if (m_starterProxyPort != 0) {
-		m_starterServer = Server::create(m_starterEndpoint.withPort(m_instance.m_starterProxyPort), USE_PROXY);
+		m_starterServer = Server::create(m_starterEndpoint.withPort(m_instance.m_starterProxyPort), option::USE_PROXY);
 	}
 	else {
 		m_starterServer = Server::create(m_starterEndpoint);