+++ tests/cpp/src/BasicRequester.cpp
@@ -30,7 +30,7 @@ int main(int argc, char *argv[]) {
 		useProxy = (string(argv[1]) == "true");
 	}
 	if (useProxy) {
-		options |= USE_PROXY;
+		options |= option::USE_PROXY;
 	}
 
 	std::unique_ptr<ServerAndApp> starter = This::connectToStarter(options);