+++ tests/cpp/src/TestSubscriberTimeout.cpp
@@ -39,7 +39,7 @@ int main(int argc, char *argv[]) {
 		useProxy = (string(argv[2]) == "true");
 	}
 	if (useProxy) {
-		options |= USE_PROXY;
+		options |= option::USE_PROXY;
 		endpoint = "tcp://localhost:12000";
 	}