+++ tests/python/src/testcancel.py
@@ -43,7 +43,7 @@ useProxy = False if len(sys.argv) < 3 else (sys.argv[1] == "true")
 options = 0
 endpoint = "tcp://localhost:11000";
 if useProxy:
-    options |= cameopy.USE_PROXY
+    options |= cameopy.option.USE_PROXY
     endpoint = "tcp://localhost:12000";
 
 server = cameopy.Server.create(endpoint, options)