+++ tests/python/src/testsimple.py
@@ -10,7 +10,7 @@ useProxy = False if len(sys.argv) < 4 else (sys.argv[2] == "true")
 options = 0
 endpoint = "tcp://localhost:11000";
 if useProxy:
-    options |= cameopy.USE_PROXY
+    options |= cameopy.option.USE_PROXY
     endpoint = "tcp://localhost:12000";
 
 server = cameopy.Server.create(endpoint, options)